+++ tests/testthat/test_GIRF.R
@@ -19,10 +19,10 @@ girf2 <- GIRF(mod222s, variables=2, shock_size=1, N=1, R2=1, R1=1, seeds=1,
               include_mixweights=FALSE, init_values=mod222s$data, ci=0.1)
 
 test_that("GIRF works correctly", {
-  expect_equal(unname(girf1$girf_res[[1]]$point_est[3,]), c(2.2587387, 2.2018105, -0.1074759, 0.1074759), tolerance=1e-4)
-  expect_equal(unname(girf1$girf_res[[2]]$point_est[1,]), c(-1.041960, 6.685912, 0.000000, 0.000000), tolerance=1e-4)
-  expect_equal(unname(girf1$girf_res[[2]]$conf_ints[3, , 2]), c(8.559077, 8.689836, 10.084591, 10.215349), tolerance=1e-4)
-  expect_equal(unname(girf1$girf_res[[1]]$conf_ints[1, 1, ]), c(1.498949, 1.191855, 0.000000, 0.000000), tolerance=1e-4)
+  expect_equal(unname(girf1$girf_res[[1]]$point_est[3,]), c(2.7647446, 2.6555342, -0.2440685, 0.2440685), tolerance=1e-4)
+  expect_equal(unname(girf1$girf_res[[2]]$point_est[1,]), c(-0.358230, 2.298642, 0.000000, 0.000000), tolerance=1e-4)
+  expect_equal(unname(girf1$girf_res[[2]]$conf_ints[3, , 2]), c(2.728152, 2.788786, 3.435544, 3.496178), tolerance=1e-4)
+  expect_equal(unname(girf1$girf_res[[1]]$conf_ints[1, 1, ]), c(1.409993, 1.121124, 0.000000, 0.000000), tolerance=1e-4)
 
   expect_equal(unname(girf2$girf_res[[1]]$point_est[1,]), c(0.1317841, -0.8456145), tolerance=1e-4)
   expect_equal(unname(girf2$girf_res[[1]]$point_est[2,]), c(-3.333572, -1.151782), tolerance=1e-4)